move document extensions
