+++ packages/espressocash_app/lib/features/kyc_sharing/screens/document_input_screen.dart
@@ -1,5 +1,3 @@
-// ignore_for_file: avoid-recursive-calls
-
 import 'dart:io';
 
 import 'package:flutter/material.dart';
@@ -13,6 +11,7 @@ import '../../../ui/loader.dart';
 import '../../../ui/snackbar.dart';
 import '../../country_picker/models/country.dart';
 import '../models/document_type.dart';
+import '../models/extensions.dart';
 import '../screens/document_camera_screen.dart';
 import '../screens/document_info_screen.dart';
 import '../services/kyc_service.dart';
@@ -472,83 +471,4 @@ class _RequiredCountryNotice extends StatelessWidget {
           ],
         ),
       );
-}
-
-enum RequirementRelationship {
-  and,
-  or,
-}
-
-extension on List<Requirement> {
-  String? parseCountryCode() {
-    for (final req in this) {
-      if (req is CountryCodeRequirement) {
-        return req.code;
-      } else if (req is AndRequirement) {
-        final code = req.requirements.parseCountryCode();
-        if (code != null) return code;
-      } else if (req is OrRequirement) {
-        final code = req.requirements.parseCountryCode();
-        if (code != null) return code;
-      }
-    }
-
-    return null;
-  }
-
-  List<DocumentType> parseDocumentTypes() {
-    final List<DocumentType> types = [];
-
-    for (final req in this) {
-      if (req is DocumentTypeRequirement) {
-        final docType = req.type.toDocumentType();
-        if (docType != null) types.add(docType);
-      } else if (req is AndRequirement) {
-        types.addAll(req.requirements.parseDocumentTypes());
-      } else if (req is OrRequirement) {
-        types.addAll(req.requirements.parseDocumentTypes());
-      }
-    }
-
-    return types;
-  }
-
-  List<DocumentField> parseRequiredFields() {
-    final List<DocumentField> fields = [];
-
-    for (final req in this) {
-      if (req is DocumentFieldRequirement) {
-        fields.add(req.field);
-      } else if (req is AndRequirement) {
-        fields.addAll(req.requirements.parseRequiredFields());
-      } else if (req is OrRequirement) {
-        fields.addAll(req.requirements.parseRequiredFields());
-      }
-    }
-
-    return fields;
-  }
-
-  RequirementRelationship determineDocumentFieldsRelationship() {
-    for (final req in this) {
-      if (req is AndRequirement &&
-          req.requirements.any((r) => r is DocumentFieldRequirement)) {
-        return RequirementRelationship.and;
-      } else if (req is AndRequirement) {
-        final relationship =
-            req.requirements.determineDocumentFieldsRelationship();
-        if (relationship == RequirementRelationship.and) {
-          return RequirementRelationship.and;
-        }
-      } else if (req is OrRequirement) {
-        final relationship =
-            req.requirements.determineDocumentFieldsRelationship();
-        if (relationship == RequirementRelationship.and) {
-          return RequirementRelationship.and;
-        }
-      }
-    }
-
-    return RequirementRelationship.or;
-  }
 }
